+++ changelog.md
@@ -1,5 +1,9 @@
 # CHANGELOG
+## 2.1.16
 
+- Bug fix [[BUG] Reset door and fog bug with token attacher ](https://github.com/p4535992/foundryvtt-arms-reach/issues/39)
+- Bug fix [[BUG] Close Open Doors/ Reset Doors and Fog doesn't work if doors are open](https://github.com/p4535992/foundryvtt-arms-reach/issues/33)
+- Bug fix [[BUG] "reset doors and fog" feature, the doors glitch and disappear.](https://github.com/p4535992/foundryvtt-arms-reach/issues/31)
 ## 2.1.15
 
 - Add Walls Reach